backend/requirements.txt
autopep8==1.5.7
pylint==2.8.2
pylint-django==2.4.4
Django==3.2.2
djangorestframework==3.12.4
django-filter==2.4.0
django-extensions==3.1.3
pyserial==3.5